move dashboard page to project section
